refactor(deferred): drop conflict check; get returns Result; extract_witness as method
Three targeted ergonomics + correctness clean-ups:

- `DeferredState::get(&Digest)` now returns `Result<&Node, SchemaError>`
  instead of `Option<&Node>`. Schema impls drop their
  `.ok_or(SchemaError::MissingNode)` and use bare `?`.
- Re-registering an expression node is silently idempotent — the
  staging-then-commit conflict-detection logic in `apply()` is gone.
  Content-addressing means same content = same digest = same node;
  a hash collision is astronomically unlikely and we don't need to
  defend against it here. Two tests that pinned the conflict
  behavior (`conflicting_insert_against_existing_errors`,
  `conflicting_insert_within_batch_errors_atomically`) are deleted.
  Assertion nodes still duplicate in the vec each time they're
  registered, which is the intended semantics.
- `extract_witness` is now a method on `DeferredState`. The free
  function and `processor/src/deferred/witness.rs` are gone; tests
  move into a `witness_tests` sub-module inside `state.rs`.

Net -61 lines. 1553 lib + 27 deferred + 3 e2e + assembly smoke pass,
clippy clean.
