+++ processor/src/deferred/mod.rs
@@ -10,7 +10,6 @@ mod handlers;
 mod schema;
 mod state;
 mod transaction;
-mod witness;
 
 pub use miden_core::deferred::{
     DeferredError, DeferredWitness, Digest, Node, Payload, Tag, hash_node,
@@ -21,4 +20,3 @@ pub use handlers::{FIELD0_ADD, FIELD0_ASSERT_EQ, FIELD0_LEAF, FIELD0_MUL, Field0
 pub use schema::{NodeType, NoopSchema, Schema, SchemaError};
 pub use state::DeferredState;
 pub use transaction::{DeferredMutation, HandlerTransaction, VmMutation};
-pub use witness::extract_witness;
